+++ src/ui/Context.h
@@ -2,10 +2,13 @@
 
 class Context
 {
+    static bool enableViewports;
+    static bool reinitImGuiNextFrame;
 public:
     static bool visible;
     static bool passThroughMouse;
     static bool inited;
+    static bool imguiInited;
 
     static void update();
     static void install_hooks();
@@ -17,4 +20,6 @@ public:
 
     static void init_imgui();
     static void deinit_imgui();
+
+    static void set_enable_viewports(bool enable);
 };