+++ src/ui/ToolBar.cpp
@@ -87,8 +87,10 @@ void ToolBar::Render() {
 				Desktop::instance->SwitchToOperationMode<ui::operation_modes::modes::level_editor::LevelEditor>();
 			if (ImGui::MenuItem("FxCol Editor"))
 				Desktop::instance->SwitchToOperationMode<ui::operation_modes::modes::fxcol_editor::FxColEditor>();
+#ifdef DEVTOOLS_TARGET_SDK_rangers
 			if (ImGui::MenuItem("SurfRide Editor"))
 				Desktop::instance->SwitchToOperationMode<ui::operation_modes::modes::surfride_editor::SurfRideEditor>();
+#endif
 			ImGui::EndMenu();
 		}