+++ src/ui/operation-modes/behaviors/Clipboard.h
@@ -12,7 +12,41 @@ private:
 	csl::ut::MoveArray<ClipboardEntry<OpModeContext>> clipboard{ GetAllocator() };
 	Traits traits;
 
+	void HandleCut() {
+		HandleCopy();
+
+		auto* selectionBehavior = operationMode.GetBehavior<SelectionBehavior<OpModeContext>>();
+		csl::ut::MoveArray<typename SelectionBehaviorTraits<OpModeContext>::ObjectType> oldObjects{ GetAllocator() };
+
+		for (auto object : selectionBehavior->GetSelection())
+			oldObjects.push_back(object);
+
+		selectionBehavior->DeselectAll();
+
+		traits.DeleteObjects(oldObjects);
+	}
+
+	void HandleCopy() {
+		Clear();
+
+		for (auto object : operationMode.GetBehavior<SelectionBehavior<OpModeContext>>()->GetSelection())
+			clipboard.push_back(traits.CreateClipboardEntry(object));
+	}
+
+	void HandlePaste() {
+		if (!traits.CanPlace())
+			return;
+
+		csl::ut::MoveArray<typename SelectionBehaviorTraits<OpModeContext>::ObjectType> newObjects{ GetAllocator() };
+
+		for (auto& entry : clipboard)
+			newObjects.push_back(traits.CreateObject(entry));
+
+		operationMode.GetBehavior<SelectionBehavior<OpModeContext>>()->Select(newObjects);
+	}
+
 public:
+	using CutAction = Action<ActionId::CUT>;
 	using CopyAction = Action<ActionId::COPY>;
 	using PasteAction = Action<ActionId::PASTE>;
 
@@ -26,33 +60,22 @@ public:
 	virtual unsigned int GetId() override { return id; }
 
 	virtual void Init() override {
+		Desktop::instance->BindShortcut<CutAction>(ShortcutId::CUT);
 		Desktop::instance->BindShortcut<CopyAction>(ShortcutId::COPY);
 		Desktop::instance->BindShortcut<PasteAction>(ShortcutId::PASTE);
 	}
 
 	virtual void Deinit() override {
+		Desktop::instance->UnbindShortcut(ShortcutId::CUT);
 		Desktop::instance->UnbindShortcut(ShortcutId::COPY);
 		Desktop::instance->UnbindShortcut(ShortcutId::PASTE);
 	}
 
 	virtual void ProcessAction(const ActionBase& action) override {
 		switch (action.id) {
-		case CopyAction::id:
-			Clear();
-
-			for (auto object : operationMode.GetBehavior<SelectionBehavior<OpModeContext>>()->GetSelection())
-				clipboard.push_back(traits.CreateClipboardEntry(object));
-
-			break;
-		case PasteAction::id: {
-			csl::ut::MoveArray<typename SelectionBehaviorTraits<OpModeContext>::ObjectType> newObjects{ hh::fnd::MemoryRouter::GetTempAllocator() };
-
-			for (auto& entry : clipboard)
-				newObjects.push_back(traits.CreateObject(entry));
-
-			operationMode.GetBehavior<SelectionBehavior<OpModeContext>>()->Select(newObjects);
-			break;
-		}
+		case CutAction::id: HandleCut(); break;
+		case CopyAction::id: HandleCopy(); break;
+		case PasteAction::id: HandlePaste(); break;
 		}
 	}